(maint) Move AppRestart facility
