+++ src/Lanceur.Infra.Win32/Restart/DummyAppRestart.cs
@@ -1,0 +1,11 @@
+namespace Lanceur.Infra.Win32.Restart;
+
+public class DummyAppRestart : IAppRestart
+{
+    #region Methods
+
+    public void Restart()
+    { }
+
+    #endregion Methods
+}
