+++ src/Lanceur.Infra.Win32/Restart/IAppRestart.cs
@@ -1,0 +1,10 @@
+namespace Lanceur.Infra.Win32.Restart;
+
+public interface IAppRestart
+{
+    #region Methods
+
+    void Restart();
+
+    #endregion Methods
+}
