+++ src/Lanceur/Views/PluginsViewModel.cs
@@ -16,6 +16,7 @@ using System.Linq;
 using System.Reactive;
 using System.Reactive.Linq;
 using System.Threading.Tasks;
+using Lanceur.Infra.Win32.Restart;
 
 namespace Lanceur.Views;
